Performance improvements fix memmove
I found the bug that was causing issues with memmove and have
added that alias back.

There is also a known performance issue on the Cortex-A72 when
using pre and post indexing. The register file has a bottle neck
of 192 bits per cycle, and each of these consumes 64 bits. From
testing this seems to mostly effect ldp's. So we instead we handle
src / srcin as a separate add/sub operation which can be optimized
by the core.

Signed-off-by: Jon Nettleton <jon@solid-run.com>